integritás ellenörző

+++ backend/src/main/java/com/starbuck/moneytracker/dto/TransactionCreateRequest.java
@@ -1,5 +1,6 @@
 package com.starbuck.moneytracker.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import org.hibernate.validator.constraints.Length;
@@ -9,17 +10,16 @@ import com.starbuck.moneytracker.entity.TransactionType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record TransactionCreateRequest (
-    @NotBlank(message = "Name is mandatory")
-    @Length(min = 3, max = 50, message = "Name must be between 3 and 50 characters")
-    String name,
+public record TransactionCreateRequest(
+        @NotBlank(message = "Name is mandatory") @Length(min = 3, max = 50, message = "Name must be between 3 and 50 characters") 
+        String name,
 
-    @NotNull(message = "transactionType is mandatory")
-    TransactionType transactionType,
+        @NotNull(message = "transactionType is mandatory") 
+        TransactionType transactionType,
 
-    @NotNull(message = "Price is mandatory")
-    Double price,
+        @NotNull(message = "Price is mandatory") 
+        BigDecimal price,
 
-    @NotNull(message = "Date is mandatory")
-    LocalDate transactionDate
-) {}
+        @NotNull(message = "Date is mandatory") LocalDate 
+        transactionDate) {
+}
