+++ backend/src/main/java/com/starbuck/moneytracker/dto/TransactionDto.java
@@ -1,13 +1,14 @@
 package com.starbuck.moneytracker.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import com.starbuck.moneytracker.entity.TransactionType;
 
 public record TransactionDto(
-        Long id,
-        String name,
-        double priceSum,
-        LocalDate transactionDate,
-        TransactionType transactionType) {
+                Long id,
+                String name,
+                BigDecimal priceSum,
+                LocalDate transactionDate,
+                TransactionType transactionType) {
 }
